+++ src/components/AllotmentLeft.tsx
@@ -4,7 +4,6 @@ import { useContext } from "react";
 import { MapContext } from "@/context/context";
 import * as rd from "@duckdb/react-duckdb";
 import { renderMapData } from "@/utils/mapFunctions";
-import { Dialog, Transition } from "@headlessui/react";
 
 import {
   Tabs,
@@ -20,6 +19,7 @@ import {
   Button,
 } from "@material-tailwind/react";
 import { RenderLayer } from "@/utils/mapFunctions";
+import { ModalAddLayer } from "@/components/ModalAddLayer";
 
 function LayerCard(props) {
   const { layer } = props;
@@ -51,15 +51,6 @@ const Layers = () => {
     { name: "ent", type: "polygon", tableName: "ent" },
   ]);
   const [collapsed, setCollapsed] = useState(false);
-  const [isOpen, setIsOpen] = useState(false);
-
-  function closeModal() {
-    setIsOpen(false);
-  }
-
-  function openModal() {
-    setIsOpen(true);
-  }
 
   // render layers on map
   layers.map((layer) => {
@@ -128,66 +119,7 @@ const Layers = () => {
           </Button>
         </div>
         <div class="px-4">
-          <IconButton onClick={openModal}>
-            <i className="fas fa-plus" />
-          </IconButton>
-          <Transition appear show={isOpen} as={Fragment}>
-            <div class="fixed ">
-              <Dialog as="div" className="relative z-20" onClose={closeModal}>
-                <Transition.Child
-                  as={Fragment}
-                  enter="ease-out duration-300"
-                  enterFrom="opacity-0"
-                  enterTo="opacity-100"
-                  leave="ease-in duration-200"
-                  leaveFrom="opacity-100"
-                  leaveTo="opacity-0"
-                >
-                  <div className="fixed inset-0 bg-black bg-opacity-25" />
-                </Transition.Child>
-
-                <div className="fixed inset-0 overflow-y-auto">
-                  <div className="flex min-h-full items-center justify-center p-4 text-center">
-                    <Transition.Child
-                      as={Fragment}
-                      enter="ease-out duration-300"
-                      enterFrom="opacity-0 scale-95"
-                      enterTo="opacity-100 scale-100"
-                      leave="ease-in duration-200"
-                      leaveFrom="opacity-100 scale-100"
-                      leaveTo="opacity-0 scale-95"
-                    >
-                      <Dialog.Panel className="w-full max-w-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
-                        <Dialog.Title
-                          as="h3"
-                          className="text-lg font-medium leading-6 text-gray-900"
-                        >
-                          Payment successful
-                        </Dialog.Title>
-                        <div className="mt-2">
-                          <p className="text-sm text-gray-500">
-                            Your payment has been successfully submitted. We’ve
-                            sent you an email with all of the details of your
-                            order.
-                          </p>
-                        </div>
-
-                        <div className="mt-4">
-                          <button
-                            type="button"
-                            className="inline-flex justify-center rounded-md border border-transparent bg-blue-100 px-4 py-2 text-sm font-medium text-blue-900 hover:bg-blue-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-blue-500 focus-visible:ring-offset-2"
-                            onClick={closeModal}
-                          >
-                            Got it, thanks!
-                          </button>
-                        </div>
-                      </Dialog.Panel>
-                    </Transition.Child>
-                  </div>
-                </div>
-              </Dialog>
-            </div>
-          </Transition>
+          <ModalAddLayer />
         </div>
       </div>
       <div class="py-10 pl-5">
